+++ src/main/kotlin/edu/umass/cics/ciir/irene/example/ExploreSDMStrategy.kt
@@ -236,9 +236,9 @@ class MaxMinCollectorManager(val mq: MultiExpr, val poolSize: Int, val epsilon: 
             totalHits.incrementAndGet()
 
             // Get bounds on expression.
-            val base = baseExpr.score()
-            val best = bestExpr.score()
-            val worst = worstExpr.score()
+            val base = baseExpr.score().toFloat()
+            val best = bestExpr.score().toFloat()
+            val worst = worstExpr.score().toFloat()
 
             // Throw in an epsilon here to avoid rounding/floating point errors causing loss.
             val max = base + best + epsilon
@@ -287,7 +287,7 @@ class MaxMinCollectorManager(val mq: MultiExpr, val poolSize: Int, val epsilon: 
                 }
                 val doc = mmsd.id - docBase
                 env.doc = doc
-                val delta = trueExpr.score()
+                val delta = trueExpr.score().toFloat()
                 val exactScore = mmsd.baseScore + delta
 
                 assert(exactScore >= mmsd.minScore && exactScore <= mmsd.maxScore) {
